Cholesky decomposition for Gaussian distribution function with permutation

This function computes the Cholesky decomposition of a covariance matrix Sigma and returns a list containing the permuted bounds for integration. The prioritization of the variables follows either the rule proposed in Gibson, Glasbey and Elston (1994), reorder variables to have outermost variables with smallest expected values. The alternative is the scheme proposed in Genz and Bretz (2009) that minimizes the variance of the truncated Normal variates.

cholperm(Sigma, l, u, method = c("GGE", "GB"))

Arguments

Sigma
d by d covariance matrix

l
d vector of lower bounds

u
d vector of upper bounds

method
string indicating which method to use. Default to "GGE"

Value

a list with components

L: Cholesky root

l: permuted vector of lower bounds

u: permuted vector of upper bounds

perm: vector of integers with ordering of permutation

Details

The list contains an integer vector perm with the indices of the permutation, which is such that Sigma(perm, perm) == L %*% t(L). The permutation scheme is described in Genz and Bretz (2009) in Section 4.1.3, p.37.

References

Genz, A. and Bretz, F. (2009). Computations of Multivariate Normal and t Probabilities, volume 105. Springer, Dordrecht.

Gibson G.J., Glasbey C.A. and D.A. Elton (1994). Monte Carlo evaluation of multivariate normal integrals and sensitivity to variate ordering. In: Dimon et al., Advances in Numerical Methods and Applications, WSP, pp. 120-126.