Women wage dataset from Mroz (1987)

The data are from the Panel Study of Income Dynamics (PSID) longitudinal study, 1976 wave. They give the number of work hours of married women along with socio-economic variables and the number of children.

Format

a data frame containing the following variables:

whrs: hours of work

kidslt6: number of children aged 5 and below years old in household

kidsge6: number of children between age of 6 and 18 in household

age: age (in years)

educ: number of years in school

hearn: hourly earnings

exp: years of previous labor market experience

Source

W. Greene's website, accessed 17.12.2019 at <http://www.stern.nyu.edu/~wgreene/Text/Edition7/TableF5-1.csv>.

References

T. A. Mroz, 1987. The Sensitivity of an Empirical Model of Married Women's Hours of Work to Economic and Statistical Assumptions, Econometrica, 55(4), pp. 765-799

See also