Normal quantile function (high precision)

Computes with tail-precision the quantile function of the standard normal distribution at \(0\le p\le 1\), and truncated to the interval \([l,u]\). Infinite values for vectors \(l\) and \(u\) are accepted.

norminvp(p, l, u)

Arguments

p
quantile at \(0\le p\le 1\)

l
lower truncation limit

u
upper truncation limit

Value

quantile value of the truncated normal distribution.

Details

Suppose we wish to simulate a random variable \(Z\) drawn from \(N(\mu,\sigma^2)\) and conditional on \(l<Z<u\) using the inverse transform method. To achieve this, first compute X=norminvp(runif(1),(l-mu)/sig,(u-mu)/sig) and then set Z=mu+sig*X

Note

If you wish to simulate truncated normal variables fast, use trandn. Using norminvp is advisable only when needed, for example, in quasi-Monte Carlo or antithetic sampling, where the inverse transform method is unavoidable.

References

Z. I. Botev (2017), The Normal Law Under Linear Restrictions: Simulation and Estimation via Minimax Tilting, Journal of the Royal Statistical Society, Series B, 79 (1), pp. 1--24.

See also

Examples

 d <- 150 # simulate via inverse transform method
 norminvp(runif(d),l = 1:d, u = rep(Inf, d))

#>   [1]   1.239325   2.145635   3.195087   4.232182   5.177946   6.008622
#>   [7]   7.125867   8.089914   9.005266  10.151198  11.179600  12.092484
#>  [13]  13.097703  14.004848  15.179541  16.239360  17.045480  18.017306
#>  [19]  19.046163  20.042015  21.066701  22.116731  23.059871  24.007536
#>  [25]  25.021619  26.016338  27.195565  28.020830  29.018183  30.010603
#>  [31]  31.018753  32.019947  33.051569  34.080323  35.074471  36.040116
#>  [37]  37.008905  38.010926  39.010080  40.008549  41.008618  42.026390
#>  [43]  43.007496  44.021501  45.001474  46.001800  47.004689  48.081830
#>  [49]  49.031898  50.062316  51.008946  52.009695  53.091779  54.045479
#>  [55]  55.015834  56.001614  57.023583  58.003331  59.044775  60.047625
#>  [61]  61.010221  62.006322  63.001905  64.013209  65.017311  66.015237
#>  [67]  67.036945  68.025211  69.033106  70.007017  71.003062  72.004841
#>  [73]  73.004077  74.002131  75.000980  76.003637  77.028040  78.005539
#>  [79]  79.001349  80.003963  81.008683  82.008774  83.045387  84.003416
#>  [85]  85.020350  86.001021  87.007211  88.001328  89.003714  90.004560
#>  [91]  91.004838  92.020754  93.018143  94.052623  95.003072  96.016290
#>  [97]  97.025941  98.012550  99.038466 100.012724 101.001875 102.021145
#> [103] 103.000559 104.006875 105.003576 106.020734 107.022659 108.010763
#> [109] 109.008735 110.013786 111.009183 112.008536 113.002518 114.010154
#> [115] 115.031866 116.001876 117.010167 118.004149 119.010621 120.000912
#> [121] 121.005810 122.011430 123.008168 124.009592 125.008780 126.010021
#> [127] 127.011453 128.003540 129.004299 130.018330 131.019987 132.014165
#> [133] 133.002055 134.020375 135.013324 136.001604 137.005397 138.014839
#> [139] 139.034453 140.011572 141.003645 142.004892 143.001175 144.000952
#> [145] 145.012556 146.000279 147.001709 148.010606 149.002378 150.001824