Truncated student generator for Bayesian regression simulation

Simulates n random vectors \(X\) exactly distributed from the d-dimensional Student distribution with df=\(\nu\) degrees of freedom, mean zero and scale matrix sigma, conditional on \(l<X<u\),

tregress(n, lb, ub, sigma, df)

Arguments

n
number of observations

lb
vector of lower truncation limits

ub
vector of upper truncation limits

sigma
scale matrix

df
degrees of freedom

Value

list with components

R: n vector of scale

Z: a d by n matrix

so that \(\sqrt(\nu)Z/R\) follows a truncated Student distribution

References

Z. I. Botev and P. L'Ecuyer (2015), Efficient probability estimation and simulation of the truncated multivariate Student-t distribution, Proceedings of the 2015 Winter Simulation Conference, pp. 380-391,

Examples

d <- 5
tregress(lb =rep(-2, d), ub = rep(2, d), df = 3, n = 10,
  sigma = diag(0.5, d) + matrix(1, d, d))

#> $R
#>  [1] 1.091621 1.985454 1.924308 2.038754 1.591299 1.729790 1.765749 1.735859
#>  [9] 2.729171 1.778877
#> 
#> $Z
#>             [,1]       [,2]        [,3]        [,4]       [,5]
#>  [1,] -0.6758494 -0.4119034 -1.01023148  0.04240809  0.6892162
#>  [2,] -1.2148716 -2.1112704 -1.33718004 -1.91212867 -1.5162984
#>  [3,] -0.4317419 -0.4487589  0.35115041 -0.64842553 -0.2188923
#>  [4,]  0.2261569 -0.2999287 -0.08212145  0.57409001  1.5177097
#>  [5,]  1.2439582 -1.0904283 -1.44418026 -1.45235922 -0.2535841
#>  [6,] -0.9776324  0.1335521  1.04890677 -1.65683327  1.0631266
#>  [7,] -1.1206467  1.5654014 -0.21995049  0.44254103  1.4139640
#>  [8,] -0.3793248 -1.1247627  0.82225738 -0.22067044  0.9058845
#>  [9,]  0.4965089 -0.4627494  0.12161087 -0.63571020  0.8978638
#> [10,]  0.1621694  0.8973604  1.17407636  0.58679919 -0.8615315
#>