Truncated student generator for Bayesian regression simulation

Simulates n random vectors \(X\) exactly distributed from the d-dimensional Student distribution with df=\(\nu\) degrees of freedom, mean zero and scale matrix sigma, conditional on \(l<X<u\),

tregress(n, lb, ub, sigma, df)

Arguments

n
number of observations

lb
vector of lower truncation limits

ub
vector of upper truncation limits

sigma
scale matrix

df
degrees of freedom

Value

list with components

R: n vector of scale

Z: a d by n matrix

so that \(\sqrt(\nu)Z/R\) follows a truncated Student distribution

References

Z. I. Botev and P. L'Ecuyer (2015), Efficient probability estimation and simulation of the truncated multivariate Student-t distribution, Proceedings of the 2015 Winter Simulation Conference, pp. 380-391,

Examples

d <- 5
tregress(lb =rep(-2, d), ub = rep(2, d), df = 3, n = 10,
  sigma = diag(0.5, d) + matrix(1, d, d))

#> $R
#>  [1] 1.071612 2.344088 1.270925 1.808098 2.639129 2.498977 3.356397 1.932050
#>  [9] 2.348136 1.527008
#> 
#> $Z
#>              [,1]       [,2]        [,3]        [,4]        [,5]
#>  [1,] -1.17654719 -1.1040449  0.47430205  0.03901914 -0.44781399
#>  [2,] -0.54471964 -0.6695404 -0.49242726 -1.05774429 -0.57148252
#>  [3,] -0.15231088  0.1927448 -0.58268282  1.36553250 -0.55947736
#>  [4,]  0.01145336  1.1024657 -0.07931285  0.69766128  1.80972542
#>  [5,]  0.56833660  0.2521940 -0.27675694 -0.21858568 -0.38443974
#>  [6,] -1.01069117 -0.9707488 -1.61277213 -0.56507086 -0.86294295
#>  [7,]  1.30670236  0.6715204  1.52583586  2.69133045  0.36763386
#>  [8,]  0.22642401 -0.5874128 -2.08483654 -1.38759995 -0.03207499
#>  [9,]  1.55534299  0.3327452  0.28419578  0.16555135  0.87146229
#> [10,] -0.37318742  1.1344740 -0.06273786 -1.55176737 -1.06284311
#>